requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
blinker==1.6.3
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
Flask==3.0.0
frozenlist==1.4.0
gunicorn==21.2.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
libsql-client==0.3.0
MarkupSafe==2.1.3
multidict==6.0.4
packaging==23.2
python-dotenv==1.0.0
typing_extensions==4.8.0
Werkzeug==3.0.1
yarl==1.9.2

pytest~=7.4.3